fix:pubsubの処理をメイン処理にする
